Fix 6 issues from second-pass audit
- scoring.ts: type recalculatePredictions param as SupabaseClient<Database>
  instead of any — RPC call now fully type-checked against database.ts
- scoring.ts: setMatchLive now detects silent RLS block via .select().single()
  (matches pattern already used in updateMatchResult)
- league-actions.tsx: LeaveLeagueButton uses try/finally so loading state
  always resets even if leaveLeague throws unexpectedly
- leaderboard-fn.sql: cast bigint aggregates to integer so PostgREST returns
  JS numbers, not strings — matches the number type in database.ts
- league-predictions.tsx: date grouping key uses toLocaleDateString("en-CA")
  for stable YYYY-MM-DD format, consistent with local-day-groups.tsx fix

Remaining known issue (not fixed — architectural change required):
- invite_code is selectable by any authenticated user via Supabase REST API
  due to the permissive leagues SELECT policy in fix-rls-invite-delete.sql.
  App-level gating (InviteSection only shown to members) prevents UI leakage
  but the DB column is still accessible via direct API calls.

+++ supabase/leaderboard-fn.sql
@@ -22,9 +22,9 @@ as $$
     p.id as user_id,
     p.username,
     p.avatar_url,
-    coalesce(sum(pr.points), 0) as total_points,
-    count(pr.points) filter (where pr.points = 5) as exact_scores,
-    count(pr.points) filter (where pr.points = 3) as correct_results,
+    coalesce(sum(pr.points), 0)::integer as total_points,
+    count(pr.points) filter (where pr.points = 5)::integer as exact_scores,
+    count(pr.points) filter (where pr.points = 3)::integer as correct_results,
     p.created_at as member_since
   from public.profiles p
   left join public.predictions pr on pr.user_id = p.id
@@ -61,9 +61,9 @@ as $$
     p.id as user_id,
     p.username,
     p.avatar_url,
-    coalesce(sum(filtered_pr.points), 0) as total_points,
-    count(filtered_pr.points) filter (where filtered_pr.points = 5) as exact_scores,
-    count(filtered_pr.points) filter (where filtered_pr.points = 3) as correct_results,
+    coalesce(sum(filtered_pr.points), 0)::integer as total_points,
+    count(filtered_pr.points) filter (where filtered_pr.points = 5)::integer as exact_scores,
+    count(filtered_pr.points) filter (where filtered_pr.points = 3)::integer as correct_results,
     lm.joined_at
   from public.league_members lm
   join public.profiles p on p.id = lm.user_id
